+++ business-logic/src/main/resources/application.properties
@@ -14,3 +14,6 @@ spring.datasource.dbcp2.default-auto-commit = true
 # ===============================
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.current_session_context_class=org.springframework.orm.hibernate5.SpringSessionContext
+
+jwt.clientId=55e18161aa7287659d9a
+jwt.client-secret=dcfd3bac80f3f893aa9acab5274cf63e5a7302cd
