+++ web/pom.xml
@@ -29,6 +29,10 @@
             <scope>compile</scope>
         </dependency>
 
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -51,7 +55,6 @@
 
 
     </dependencies>
-
 
 
 </project>
