+++ web/src/main/java/com/epam/esm/security/WebSecurityConfig.java
@@ -1,5 +1,6 @@
 package com.epam.esm.security;
 
+import com.epam.esm.handler.JwtAuthenticationEntryPoint;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -30,6 +31,8 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 
     @Autowired
     private JwtAuthenticationEntryPoint unauthorizedHandler;
+
+
 
     @Override
     @Bean
